Poetry: Clarify license and use SPDX identifiers
Without that fix, it is not clear if the license is the GPL-2.0-only
or the GPL-2.0-or-later. This has an impact for users of the library
as a project licensed under the GPLv3 can be combined with code under
the GPLv2 or later but not with code under the GPLv2 only.

While the license field is optional in poetry[1], both the poetry
documentation[1] and the PEP 621[2] strongly advise to use SPDX
license identifier.

[1]https://python-poetry.org/docs/pyproject/#license
[2]https://peps.python.org/pep-0621/#license

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>